py-ws/requirements.txt
Pillow==10.0.0
websockets==11.0.3
